# doc/api-swagger.yml
swagger: '2.0'
info:
  title: SDG backend API
  description: This API allow to store project configuration and provide a token trading system based on project participation
  version: 1.0.0
paths:
...
  '/api/v1/user/{username}/promote':
    patch:
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        default:
          description: ''
  '/api/v1/user/{username}/demote':
    patch:
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        default:
          description: ''
...